frontend and backend update push
This push integrates a huge frontend and backend prototype.
It is quite heavy, containing reworked Python code; a full frontend
prototype using Thymeleaf, html, css and js; a backend using SpringBoot,
integrated with a locally started Python Server, Gemini API requests and
communication with the Qdrant DB. The server-client connection uses sse
to ensure the frontend is up to date with the backends shenanigans.